+++ dsmr_frontend/migrations/0003_drop_history_page.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('dsmr_frontend', '0002_recent_history_weeks'),
+    ]
+
+    operations = [
+        migrations.RemoveField(
+            model_name='frontendsettings',
+            name='recent_history_weeks',
+        ),
+    ]