+++ dsmr_frontend/urls.py
@@ -3,7 +3,6 @@ from django.conf import settings
 from django.conf.urls import url
 
 from dsmr_frontend.views.dashboard import Dashboard
-from dsmr_frontend.views.history import History
 from dsmr_frontend.views.archive import Archive, ArchiveXhrSummary, ArchiveXhrGraphs
 from dsmr_frontend.views.statistics import Statistics
 from dsmr_frontend.views.trends import Trends
@@ -13,11 +12,6 @@ from dsmr_frontend.views.configuration import Configuration
 
 urlpatterns = [
     url(r'^$', Dashboard.as_view(), name='dashboard'),
-    url(
-        r'^history$',
-        cache_page(settings.CACHES['default']['TIMEOUT'])(History.as_view()),
-        name='history'
-    ),
     url(r'^archive$', Archive.as_view(), name='archive'),
     url(r'^archive/xhr/summary$', ArchiveXhrSummary.as_view(), name='archive-xhr-summary'),
     url(r'^archive/xhr/graphs$', ArchiveXhrGraphs.as_view(), name='archive-xhr-graphs'),