+++ dsmrreader/__init__.py
@@ -1,6 +1,6 @@
 from django.utils.version import get_version
 
 
-VERSION = (1, 0, 0, 'rc', 2)
+VERSION = (1, 0, 0, 'rc', 3)
 
 __version__ = get_version(VERSION)